requirements.txt
fastapi
nicegui
uvicorn
httpx
dotenv
pymongo
yfinance
alpaca-py
<<<<<<< Updated upstream
tqdm
=======
plotly
apscheduler
>>>>>>> Stashed changes
